% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.r
\docType{methods}
\name{arithmetic}
\alias{*,BandMat,BandMat-method}
\alias{*,BandMat,numeric-method}
\alias{*,numeric,BandMat-method}
\alias{arithmetic}
\title{Arithmetic Operators}
\usage{
\S4method{*}{BandMat,BandMat}(e1, e2)

\S4method{*}{BandMat,numeric}(e1, e2)

\S4method{*}{numeric,BandMat}(e1, e2)
}
\arguments{
\item{e1, e2}{Banded matrices or numeric vectors}
}
\value{
A banded matrix.
}
\description{
Binary operations for banded/banded, banded/matrix, and banded/vector
operations.
}
\details{
If \code{e1} and \code{e2} are banded matrices, then they must be
conformable and packed in the same way.
}
\keyword{Methods}

