% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.banded.r
\name{as.banded}
\alias{as.banded}
\title{as.banded}
\usage{
as.banded(x, kl, ku)
}
\arguments{
\item{x}{A matrix containing numeric, integer, or logical data.}

\item{kl, ku}{The size of the lower and upper bands.  At least one value \code{kl} or
\code{ku} must be supplied; however, if one is missing, the other will
be assumed to be the same value. Each supplied value should be a
non-negative integer. See the details section for more information
about these values.}
}
\value{
A banded matrix.
}
\description{
Convert a regular R matrix into a banded matrix.
}
\details{
The values \code{kl} and \code{ku} describe the bandwidth of the band
matrix.  If they are different, they will produce a general banded
matrix.  If they are both 0, they will produce a diagonal matris.
}
\examples{
x <- matrix(1:25, nrow=5)

# A diagonal matrix
as.banded(x, 0)

# A general banded matrix
as.banded(x, 2, 1)

}
\references{
\url{http://www.netlib.org/lapack/lug/node124.html}
}
\seealso{
\code{\link{kdim}} and \code{\link{classes}}
}
\keyword{Casters}

