% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.r
\docType{methods}
\name{as.matrix}
\alias{as.matrix}
\alias{as.matrix,BandMat-method}
\alias{as.matrix,SymMat-method}
\title{as.matrix}
\usage{
\S4method{as.matrix}{BandMat}(x)

\S4method{as.matrix}{SymMat}(x)
}
\arguments{
\item{x}{A banded or symmetric matrix.}
}
\value{
An R matrix.
}
\description{
as.matrix
}
\keyword{Casters}

