% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.symmetric.r
\name{as.symmetric}
\alias{as.symmetric}
\title{as.symmetric}
\usage{
as.symmetric(x, triangle = "u")
}
\arguments{
\item{x}{A matrix containing numeric, integer, or logical data.}

\item{kl, ku}{The size of the lower and upper bands.  At least one value \code{kl} or
\code{ku} must be supplied; however, if one is missing, the other will
be assumed to be the same value. Each supplied value should be a
non-negative integer. See the details section for more information
about these values.}
}
\value{
A symmetric matrix.
}
\description{
Convert a regular R matrix into a packed symmetric matrix.
}
\details{
The values \code{kl} and \code{ku} describe the bandwidth of the band
matrix.  If they are different, they will produce a general banded
matrix.  If they are both 0, they will produce a diagonal matris.
}
\examples{
x <- matrix(1:9, nrow=3)

# A diagonal matrix
as.symmetric(x)

}
\seealso{
\code{\link{kdim}} and \code{\link{classes}}
}
\keyword{Casters}

