% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-classes.r
\docType{class}
\name{classes}
\alias{DiagMat-class}
\alias{GenBandMat-class}
\alias{SymBandMat-class}
\alias{ZeroMat-class}
\alias{classes}
\title{Banded Matrix Classes}
\description{
Class for general and symmetric banded matrices.
}
\section{Slots}{

\describe{
\item{\code{DATA}}{The data.}

\item{\code{dim}}{matrix dimension.}

\item{\code{kl,ku}}{Size of the upper and lower bands (bandwidth) in the general case.}

\item{\code{k}}{Size of the bands in the symmetric case.}
}}
\keyword{Classes}

