% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.r
\name{constructors}
\alias{constructors}
\alias{diagmat}
\alias{genbandmat}
\alias{zeromat}
\title{Banded Matrix Constructors}
\usage{
zeromat(dim)

diagmat(Data, dim)

genbandmat(Data, dim, kl, ku)
}
\arguments{
\item{dim}{The matrix dimension.}

\item{Data}{The compact band matrix representation.}

\item{kl, ku}{The lower and upper bandwidths, respectively.}
}
\description{
Constructors for banded matrices. WARNING: for advanced users only.
Use the caster function for conversions.
}
\details{
See url in the references for an explanation of the storage.
}
\references{
\url{http://www.netlib.org/lapack/lug/node124.html}
}
\seealso{
\code{\link{as.banded}} and \code{\link{as.matrix}}
}

