% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eig.r
\name{eig}
\alias{eig}
\title{eig}
\usage{
eig(x, only.values = FALSE, descending = TRUE)
}
\arguments{
\item{x}{The input matrix.}

\item{only.values}{Should only the eigenvalues be returned, or should the eigenvectors
also be returned?}

\item{descending}{Should the values/vectors be returned in descending order,
or ascending order?  By default, LAPACK returns them in ascending
order, so putting them in descending order (R's default in \code{eigen()})
has a performance cost.}
}
\value{
A matrix-like object of the same type as the input.
}
\description{
Eigenvalue decomposition.
}
\details{
Runtime performance for a SymMat (symmetric packed matrix) will be
MUCH slower than when using full (non-packed) storage.
}

