% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-accessors.r
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(x)
}
\arguments{
\item{x}{A band matrix.}
}
\description{
Extract the internal storage representation of the banded matrix. WARNING:
for advanced users only.
}
\details{
See url in the references for an explanation of the storage.
}
\references{
\url{http://www.netlib.org/lapack/lug/node124.html}
}
\keyword{Accessors}

