% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-accessors.r
\name{getdim}
\alias{getdim}
\title{getdim}
\usage{
getdim(x)
}
\arguments{
\item{x}{A band matrix.}
}
\description{
Get the dimension of the banded matrix. Like \code{dim()} for R matrices.
}
\keyword{Accessors}

