% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSym.r
\name{isSym}
\alias{isSym}
\title{isSym}
\usage{
isSym(x)
}
\arguments{
\item{x}{A matrix-like object.}
}
\value{
\code{TRUE} or \code{FALSE}, accordingly.
}
\description{
Test if a matrix-like object is symmetric.  In addition to supporting
the new classes, this method is SIGNIFICANTLY more efficient
than R's \code{isSymmetric()} for matrix types.
}
\details{
For a matrix input, we use a block algorithm.
}

