% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSymmetric.r
\docType{methods}
\name{isSymmetric}
\alias{isSymmetric}
\alias{isSymmetric,DiagMat-method}
\alias{isSymmetric,GenBandMat-method}
\alias{isSymmetric,SymMat-method}
\alias{isSymmetric,ZeroMat-method}
\title{isSymmetric}
\usage{
\S4method{isSymmetric}{DiagMat}(object, ...)

\S4method{isSymmetric}{ZeroMat}(object, ...)

\S4method{isSymmetric}{SymMat}(object, ...)

\S4method{isSymmetric}{GenBandMat}(object, ...)
}
\arguments{
\item{object}{A matrix-like object.}

\item{...}{Ignored.}
}
\value{
\code{TRUE} or \code{FALSE}, accordingly.
}
\description{
Test if a matrix is symmetric.
}

