% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print-BandMat}
\alias{print-BandMat}
\alias{print.BandMat}
\title{Print \code{BandMat} objects}
\usage{
\method{print}{BandMat}(x, ...)
}
\arguments{
\item{x}{\code{BandMat} object}

\item{...}{unused}
}
\description{
Printing for \code{BandMat} objects.
}

