% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrize.r
\name{symmetrize}
\alias{symmetrize}
\title{symmetrize}
\usage{
symmetrize(x, triangle = "u")
}
\arguments{
\item{x}{A matrix-like object.}

\item{triangle}{A single character, either "u" or "l" for upper or lower, respectively.
This determines which triangle is preserved.  The default is the upper
triangle.  Case is ignored.}
}
\value{
A matrix-like object of the same type as the input.
}
\description{
Make a matrix-like object symmetric.
}
\details{
For a matrix input, we use a block algorithm.
}

