% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.r
\docType{methods}
\name{t,BandMat-method}
\alias{t,BandMat-method}
\title{t}
\usage{
\S4method{t}{BandMat}(x)
}
\arguments{
\item{x}{A banded matrix.}
}
\value{
A banded matrix of the same type.
}
\description{
t
}
\keyword{LinAlg}

