% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.r
\docType{methods}
\name{xpose}
\alias{t,BandMat-method}
\alias{t,GenBandMat-method}
\alias{t,SymMat-method}
\alias{t,matrix-method}
\alias{xpose}
\title{t}
\usage{
\S4method{t}{BandMat}(x)

\S4method{t}{GenBandMat}(x)

\S4method{t}{SymMat}(x)

\S4method{t}{matrix}(x)
}
\arguments{
\item{x}{A banded matrix.}
}
\value{
A banded matrix of the same type.
}
\description{
t
}
\keyword{LinAlg}

