% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.banded.r
\name{as.banded}
\alias{as.banded}
\title{as.banded}
\usage{
as.banded(x, kl, ku)
}
\arguments{
\item{x}{A matrix containing numeric, integer, or logical data.}

\item{kl, ku}{The size of the lower and upper bands.  See the details section for more
information.  Each should be a non-negative integer.}
}
\value{
A banded matrix.
}
\description{
Convert a regular R matrix into a banded matrix.
}
\details{
The values \code{kl} and \code{ku} describe the bandwidth of the band
matrix.  If they are different, they will produce a general banded
matrix.  If they are both 0, they will produce a diagonal matris.
}
\seealso{
\code{\link{kdim}} and \code{\link{classes}}
}

