% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.symmetric.r
\name{as.symmetric}
\alias{as.symmetric}
\title{as.symmetric}
\usage{
as.symmetric(x, triangle = "u")
}
\arguments{
\item{x}{A matrix containing numeric, integer, or logical data.}

\item{triangle}{A single character, either "u" or "l" for upper or lower, respectively.
This determines which triangle is used to create the packed storage matrix.
The default is the upper triangle.  Case is ignored.}
}
\value{
A symmetric matrix.
}
\description{
Convert a regular R matrix into a packed symmetric matrix, whether or not
the input matrix was symmetric. Test for symmetry with R's
\code{isSymmetric()}.
}
\details{
The storage scheme is packed symmetric storage, so only the diagonal and
the specified triangle are stored.
}
\examples{
x <- matrix(1:25, nrow=5)
# use the upper triangle. note that 'x' is not symmetric, but the return is
as.symmetric(x)

# lower triangle
as.symmetric(x, 'l')

}
\references{
\url{http://www.netlib.org/lapack/lug/node123.html}
}
\seealso{
\code{\link{kdim}} and \code{\link{classes}}
}
\keyword{Casters}

