% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSym.r
\name{isSym}
\alias{isSym}
\title{isSym}
\usage{
isSym(x)
}
\arguments{
\item{object}{A matrix-like object.}

\item{...}{Ignored.}
}
\value{
\code{TRUE} or \code{FALSE}, accordingly.
}
\description{
Test if a matrix is symmetric.
}

