% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.r
\name{xpose}
\alias{xpose}
\alias{xpose.SymMat}
\title{xpose}
\usage{
xpose(x)

\method{xpose}{SymMat}(x)
}
\arguments{
\item{x}{A banded matrix.}
}
\value{
A matrix-like object of the same type as the input.
}
\description{
Transpose a matrix-like thing.
}
\keyword{LinAlg}

